Made biome ignore worker-configuration.d.ts as much as possible.
This is an auto-generated file, and biome is exceedingly unhappy with it. Following the advice on
`https://biomejs.dev/reference/configuration/#interaction-with-the-scanner` it seems we should not run the linter on it
because it is an "output" file (though we should still check it in for more reproducible deploys, it seems).