Switched from tabs to spaces.
This seems silly, but I'm pretty sure my editor (Zed) is sometimes prone to formatting things with spaces.

My guess is this is a bug with my editor, or that biome was not running at some point and my editor decided to format
things on its own (so, arguably still a bug in my editor).

Anyway I prefer spaces, if you have a strong preference the other way please just revert this commit and I'll figure out
my issue with my editor.