Basic structure of the API.
Yes, this is overcomplicated and redundant. Thankfully, there are only 2 handlers we need to implement for the MVP:
PUT for `api/events/[eventId]/availability/[userId]`and GET for `api/events/[eventId]`.

For the final project, we also need to implement POST for `src/pages/api/events` (to create an event) and POST for
`src/pages/api/users` (to create a user).

At least, that's my understanding so far.

This is my first REST API that I hope to actually implement! I'm sure it shows, but I also think it's a good starting
point.